chore(ts-sdk): Refactor rename checkpoint.ts
