+++ sdk/graphql-transport/src/mappers/checkpoint.ts
